add comment controller and model files and restructure accordingly
